// resources/js/Pages/Admin/Settings/types.ts
export type MailDriver = 'smtp' | 'sendmail';
export type MailEncryption = 'tls' | 'ssl' | 'none';

export interface EmailSettings {
    mail_mailer: MailDriver;
    smtp_host: string;
    smtp_port: number;
    smtp_username: string;
    smtp_has_password: boolean;
    smtp_encryption: MailEncryption;
    mail_from_address: string;
    mail_from_name: string;
    admin_notification_email: string;
    admin_notification_events: string[];
}